refactor(operations): harden callOp retry loop — bounded while, no magic fallback, barrel exports
- Replace while(true) with while(attempt <= MAX_COMPLETE_RETRY_ATTEMPTS=20) + NaxError on overflow
- Remove dead `?? 0` fallback on decision.delayMs (typed as number, never undefined at that point)
- Export _callOpDeps from src/operations barrel so tests use barrel not internal path
- Fix test imports to use ../../../src/operations barrel
